# server/docs/swagger.yaml
definitions:
...
  github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.Playmate:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      description:
        type: string
      game:
        type: string
      gender:
        type: string
      id:
        type: integer
      isOnline:
        type: boolean
      level:
        type: integer
      likes:
        type: integer
      nickname:
        type: string
      price:
        type: number
      rank:
        type: string
      rating:
        type: number
      tags:
        description: 用逗号分隔的标签
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      balance:
        type: number
      bio:
        type: string
      birthday:
        type: string
      couponCount:
        type: integer
      createdAt:
        type: string
      gender:
        type: string
      id:
        type: integer
      location:
        type: string
      nickname:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
      username:
        type: string
      vipLevel:
        type: integer
    type: object
  model.UserBrowseHistory:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      playmateId:
        type: integer
      userId:
        type: integer
      viewedAt:
        type: string
    type: object
  model.UserSettings:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      language:
        type: string
      notifications:
        description: JSON格式存储通知设置
        type: string
      privacy:
        description: JSON格式存储隐私设置
        type: string
      theme:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  request.RegisterRequest:
    properties:
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - phone
    - username
    type: object
...
  request.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      birthday:
        type: string
      gender:
        type: string
      location:
        type: string
      nickname:
        type: string
      phone:
        type: string
    type: object
  request.UpdateSettingsRequest:
    properties:
      language:
        type: string
      notifications:
        additionalProperties:
          type: boolean
        type: object
      privacy:
        additionalProperties:
          type: boolean
        type: object
      theme:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /playmate/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 用户登录
      tags:
      - Auth
  /playmate/auth/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户登出
      tags:
      - Auth
  /playmate/auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 刷新token
      tags:
      - Auth
  /playmate/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 用户注册
      tags:
      - Auth
...
  /playmate/user/favorites:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Playmate'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取收藏列表
      tags:
      - User
  /playmate/user/favorites/{favoriteId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 收藏ID
        in: path
        name: favoriteId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 移除收藏
      tags:
      - User
  /playmate/user/following:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Playmate'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取关注列表
      tags:
      - User
  /playmate/user/following/{userId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消关注成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 取消关注用户
      tags:
      - User
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 关注成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 关注用户
      tags:
      - User
  /playmate/user/history:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 清空成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 清空浏览历史
      tags:
      - User
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserBrowseHistory'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取浏览历史
      tags:
      - User
  /playmate/user/info:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - User
  /playmate/user/profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: 个人资料
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新个人资料
      tags:
      - User
  /playmate/user/settings:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSettings'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户设置
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: 设置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSettings'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户设置
      tags:
      - User
  /playmate/user/wallet:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取钱包信息
      tags:
      - User
  /playmate/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...